Move the AnnotationBuilder to the javacutil project.
