+++ javacutil/src/org/checkerframework/javacutil/AnnotationUtils.java
@@ -1,8 +1,6 @@
 package org.checkerframework.javacutil;
 
 /*>>>
-import org.checkerframework.dataflow.qual.Pure;
-import org.checkerframework.dataflow.qual.SideEffectFree;
 import org.checkerframework.checker.nullness.qual.*;
 import org.checkerframework.checker.interning.qual.*;
 */
@@ -28,14 +26,11 @@ import java.util.TreeMap;
 import java.util.TreeSet;
 import javax.lang.model.element.AnnotationMirror;
 import javax.lang.model.element.AnnotationValue;
-import javax.lang.model.element.ElementKind;
 import javax.lang.model.element.ExecutableElement;
 import javax.lang.model.element.Name;
 import javax.lang.model.element.TypeElement;
 import javax.lang.model.type.DeclaredType;
-import javax.lang.model.type.TypeMirror;
 import javax.lang.model.util.ElementFilter;
-import javax.lang.model.util.Elements;
 
 /** A utility class for working with annotations. */
 public class AnnotationUtils {
@@ -47,7 +42,7 @@ public class AnnotationUtils {
 
     // TODO: hack to clear out static state.
     public static void clear() {
-        annotationsFromNames.clear();
+        AnnotationBuilder.clear();
         annotationMirrorNames.clear();
         annotationMirrorSimpleNames.clear();
         annotationClassNames.clear();
@@ -56,10 +51,6 @@ public class AnnotationUtils {
     // **********************************************************************
     // Factory Methods to create instances of AnnotationMirror
     // **********************************************************************
-
-    /** Caching for annotation creation. */
-    private static final Map<CharSequence, AnnotationMirror> annotationsFromNames =
-            Collections.synchronizedMap(new HashMap<CharSequence, AnnotationMirror>());
 
     private static final int ANNOTATION_CACHE_SIZE = 500;
 
@@ -89,79 +80,6 @@ public class AnnotationUtils {
             annotationClassNames =
                     Collections.synchronizedMap(
                             new HashMap<Class<? extends Annotation>, /*@Interned*/ String>());
-
-    /**
-     * Creates an {@link AnnotationMirror} given by a particular fully-qualified name.
-     * getElementValues on the result returns an empty map.
-     *
-     * @param elements the element utilities to use
-     * @param name the name of the annotation to create
-     * @return an {@link AnnotationMirror} of type {@code} name
-     */
-    public static AnnotationMirror fromName(Elements elements, CharSequence name) {
-        AnnotationMirror res = annotationsFromNames.get(name);
-        if (res != null) {
-            return res;
-        }
-        final DeclaredType annoType = typeFromName(elements, name);
-        if (annoType == null) {
-            return null;
-        }
-        if (annoType.asElement().getKind() != ElementKind.ANNOTATION_TYPE) {
-            ErrorReporter.errorAbort(annoType + " is not an annotation");
-            return null; // dead code
-        }
-        AnnotationMirror result =
-                new AnnotationMirror() {
-                    String toString = "@" + annoType;
-
-                    @Override
-                    public DeclaredType getAnnotationType() {
-                        return annoType;
-                    }
-
-                    @Override
-                    public Map<? extends ExecutableElement, ? extends AnnotationValue>
-                            getElementValues() {
-                        return Collections.emptyMap();
-                    }
-                    /*@SideEffectFree*/
-                    @Override
-                    public String toString() {
-                        return toString;
-                    }
-                };
-        annotationsFromNames.put(name, result);
-        return result;
-    }
-
-    /**
-     * Creates an {@link AnnotationMirror} given by a particular annotation class.
-     *
-     * @param elements the element utilities to use
-     * @param clazz the annotation class
-     * @return an {@link AnnotationMirror} of type given type
-     */
-    public static AnnotationMirror fromClass(Elements elements, Class<? extends Annotation> clazz) {
-        return fromName(elements, clazz.getCanonicalName());
-    }
-
-    /**
-     * A utility method that converts a {@link CharSequence} (usually a {@link String}) into a
-     * {@link TypeMirror} named thereby.
-     *
-     * @param elements the element utilities to use
-     * @param name the name of a type
-     * @return the {@link TypeMirror} corresponding to that name
-     */
-    private static DeclaredType typeFromName(Elements elements, CharSequence name) {
-        /*@Nullable*/ TypeElement typeElt = elements.getTypeElement(name);
-        if (typeElt == null) {
-            return null;
-        }
-
-        return (DeclaredType) typeElt.asType();
-    }
 
     // **********************************************************************
     // Helper methods to handle annotations.  mainly workaround
